requirements.txt
discord==1.7.3
python-telegram-bot